Update to openjdk 8.121-b13
This change also brings the packaging in line with Solus requirements
and standards, as it seems previous contributors have used other distros
for reference first.

Additionally, we ensure we fully control the compiler environment with the
newly synced flags, so that the build is actually in line with Solus binary
requirements too.

Signed-off-by: Ikey Doherty <ikey@solus-project.com>